pyproject.toml
[tool.poetry]
name = "metroid"
version = "1.1.0"  # Remember to change in __init__.py as well
description = "Metroid - Metro for Django"
authors = ["Jonas Krüger Svensson <jonas.svensson@intility.no>"]
maintainers = [
    "Ali Arfan <ali.arfan@intility.no>",
    "Ingvald Lorentzen <ingvald.lorentzen@intility.no",
    "Per Anders Stadheim <per.anders.stadheim@intility.no>",
]
readme = "README.md"
homepage = "https://github.com/intility/metroid"
repository = "https://github.com/intility/metroid"
documentation = "https://github.com/intility/metroid"
keywords = [
    'async', 'django', 'servicebus', 'task', 'celery', 'worker', 'rq',
]
classifiers = [
    'Development Status :: 5 - Production/Stable',
    'Environment :: Web Environment',
    'Framework :: Django',
    'Framework :: Django :: 3.1',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: BSD License',
    'Operating System :: OS Independent',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3.9',
    'Topic :: Software Development',
    'Topic :: Software Development :: Libraries',
    'Topic :: Software Development :: Libraries :: Application Frameworks',
    'Topic :: Software Development :: Libraries :: Python Modules',
]

[tool.poetry.dependencies]
python = "^3.8"
azure-servicebus = "^7.0.1"
Django = "^3.1.1"
django-guid = "^3.2.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.9.3"
python-decouple = "^3.4"
redis = "^3.5.3"
black = "^20.8b1"
pytest = "^6.2.2"
pytest-cov = "^2.11.1"
pytest-django = "^4.1.0"
pytest-asyncio = "^0.14.0"
pytest-mock = "^3.5.1"
requests-mock = "^1.8.0"
pytest-freezegun = "^0.4.2"
celery = "5.0.0"
django-rq = "^2.4.1"
redislite = "^6.0.674960"

[tool.black]
line-length = 120
skip-string-normalization = true
target-version = ['py37']
include = '\.pyi?$'
exclude = '''
(
    (\.eggs|\.git|\.hg|\.mypy_cache|\.tox|\.venv|\venv|\.github|\docs|\tests|\__pycache__)
)
'''

[tool.isort]
profile = "black"
src_paths = ["metroid"]
combine_as_imports = true
line_length = 120
sections = [
    'FUTURE',
    'STDLIB',
    'DJANGO',
    'THIRDPARTY',
    'FIRSTPARTY',
    'LOCALFOLDER'
]
known_django = ['django']

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = 'demoproj.settings'
log_cli_format = '%(levelname)s %(asctime)s [%(correlation_id)s] %(name)s %(message)s'
log_cli = true
log_cli_level = 'DEBUG'


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
